Move all new authorizer finders to kg module
