+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/MicroserviceRoutes.scala
@@ -42,7 +42,7 @@ import io.renku.http.server.QueryParameterTools._
 import io.renku.http.server.security.Authentication
 import io.renku.http.server.security.model.{AuthUser, MaybeAuthUser}
 import io.renku.http.server.version
-import io.renku.knowledgegraph.datasets.{DatasetIdRecordsFinder2, DatasetSameAsRecordsFinder2}
+import io.renku.knowledgegraph.datasets.{DatasetIdRecordsFinder2, DatasetSameAsRecordsFinder2, ProjectAuthRecordsFinder}
 import io.renku.knowledgegraph.datasets.details.RequestedDataset
 import io.renku.metrics.{MetricsRegistry, RoutesMetrics}
 import io.renku.triplesstore.{ProjectSparqlClient, ProjectsConnectionConfig, SparqlQueryTimeRecorder}
