+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/datasets/DatasetSameAsRecordsFinder2.scala
@@ -20,29 +20,41 @@ package io.renku.knowledgegraph.datasets
 
 import cats.effect.kernel.Sync
 import fs2.Stream
+import eu.timepit.refined.auto._
 import io.renku.graph.http.server.security.Authorizer
 import io.renku.graph.model.datasets
 import io.renku.http.server.security.model
 import io.renku.jsonld.EntityId
 import io.renku.projectauth.util.ProjectAuthDataRow
-import io.renku.triplesstore.ProjectSparqlClient
+import io.renku.triplesstore.{ProjectSparqlClient, SparqlQueryTimeRecorder}
 import io.renku.triplesstore.client.syntax._
+import org.typelevel.log4cats.Logger
 
 trait DatasetSameAsRecordsFinder2[F[_]] extends Authorizer.SecurityRecordFinder[F, datasets.SameAs]
 
 object DatasetSameAsRecordsFinder2 {
 
-  def apply[F[_]: Sync](projectSparqlClient: ProjectSparqlClient[F]): DatasetSameAsRecordsFinder2[F] =
+  def apply[F[_]: Sync: Logger: SparqlQueryTimeRecorder](
+      projectSparqlClient: ProjectSparqlClient[F]
+  ): DatasetSameAsRecordsFinder2[F] =
     new Impl[F](projectSparqlClient)
 
-  private class Impl[F[_]: Sync](projectSparqlClient: ProjectSparqlClient[F]) extends DatasetSameAsRecordsFinder2[F] {
-    override def apply(id: datasets.SameAs, user: Option[model.AuthUser]): F[List[Authorizer.SecurityRecord]] =
+  private class Impl[F[_]: Sync: Logger: SparqlQueryTimeRecorder](projectSparqlClient: ProjectSparqlClient[F])
+      extends DatasetSameAsRecordsFinder2[F] {
+    private[this] val timeRecorder = SparqlQueryTimeRecorder[F]
+
+    override def apply(sameAs: datasets.SameAs, user: Option[model.AuthUser]): F[List[Authorizer.SecurityRecord]] =
       Stream
-        .evals(projectSparqlClient.queryDecode[ProjectAuthDataRow](query(id)))
+        .evals(runQuery(sameAs))
         .through(ProjectAuthDataRow.collect)
         .map(p => Authorizer.SecurityRecord(p.visibility, p.slug, p.members.map(_.gitLabId)))
         .compile
         .toList
+
+    private def runQuery(sameAs: datasets.SameAs) =
+      timeRecorder.reportTime("security-dataset-sameAs")(
+        projectSparqlClient.queryDecode[ProjectAuthDataRow](query(sameAs))
+      )
 
     private def query(sameAs: datasets.SameAs) =
       sparql"""PREFIX schema: <http://schema.org/>
